Elixir In Action v0.1.0 Chapter2.Calculator

Documentation for 'Calculator'.

Contains the documentation for calculator modules used in the book Elixir in Action 2nd Edition

Functions

division(a, b)

Specs

division(number(), number()) :: float()

Divides a number by the other

Book section: 2.4.1

This exercise explains that the '/' operator always returns a float

Examples

iex> Chapter2.Calculator.division(1,1)
1.0

iex> Chapter2.Calculator.division(1,4)
0.25

integer_of_div(a, b)

Specs

integer_of_div(integer(), integer()) :: integer()

Gets the integer part of a division

Book section: 2.4.1

This exercise explains how to use the kernel function div

Examples

iex> Chapter2.Calculator.integer_of_div(1,1)
1

iex> Chapter2.Calculator.integer_of_div(5,2)
2

remainder_of_div(a, b)

Specs

remainder_of_div(integer(), integer()) :: integer()

Gets the remainder part of a division

Book section: 2.4.1

This exercise explains how to use the kernel function rem

Examples

iex> Chapter2.Calculator.remainder_of_div(1,1)
0

iex> Chapter2.Calculator.remainder_of_div(1,4)
1

sum(a, b \\ 0)

Specs

sum(number(), number()) :: number()

Sum two numbers

Book section: 2.3.3

This exercise explains the usual pattern that low arity delegates to higher arity functions with some default argument values

Examples

iex> Chapter2.Calculator.sum(1)
1

iex> Chapter2.Calculator.sum(1,4)
5