Elixir In Action v0.1.0 Chapter2.Geometry

Documentation for 'Geometry'.

Contains the documentation for geometries modules used in the book Elixir in Action 2nd Edition

Functions

circle_area(r)

Specs

circle_area(number()) :: float()

Calculate the area of a circle

Book section: 2.3.6

This exercise explains the usage of module attributes and how to combine functions in docTests. In this case I'm using @pi as a module attribute

Examples

iex> Chapter2.Geometry.circle_area(3) |> Float.round(2)
28.27

circle_circunference(r)

Specs

circle_circunference(number()) :: float()

Calculate the circunference of a circle

Book section: 2.3.6

This exercise explains the usage of module attributes and how to combine functions in docTests. In this case I'm using @pi as a module attribute

Examples

iex> Chapter2.Geometry.circle_circunference(3) |> Float.round(2)
18.85

rectangle_area(a, b)

Specs

rectangle_area(non_neg_integer(), non_neg_integer()) :: non_neg_integer()

Calculate the area of a rectangle

Book section: 2.3.1

This exercise explains the creation and use of a function

Examples

iex> Chapter2.Geometry.rectangle_area(3,2)
6

iex> Chapter2.Geometry.rectangle_area(1,7)
7

square_area(a)

Specs

square_area(non_neg_integer()) :: non_neg_integer()

Calculate the area of a square

Book section: 2.3.3

This exercise explains the usual pattern that low arity delegates to higher arity functions with some default argument values

In this case, a square is a subset of rectangles so the area function can be implemented as a rectangle area call with both parameters having equal values

Examples

iex> Chapter2.Geometry.square_area(3)
9

iex> Chapter2.Geometry.square_area(7)
49