Elixir In Action v0.1.0 Chapter3.CheckNumber

This module explain how guards work on multi-clauses functions.

Book section: 3.2

Functions

check(number)

Specs

check(number()) :: :negative | :positive | :zero

Indicates if a number is positive, negative or zero.

Book section: 3.2.2

This exercises shows how to use guards on multi-clauses functions to achieve "if like" behaviours

Examples

iex> Chapter3.CheckNumber.check(2)
:positive

iex> Chapter3.CheckNumber.check(0)
:zero

iex> Chapter3.CheckNumber.check(-1)
:negative

cond_check(number)

Specs

cond_check(any()) :: :negative | :positive | :zero

Indicates if a number is positive, negative or zero.

Book section: 3.3.2

This exercises shows how to use branching with cond macro

Examples

iex> Chapter3.CheckNumber.cond_check(2)
:positive

iex> Chapter3.CheckNumber.cond_check(0)
:zero

iex> Chapter3.CheckNumber.cond_check(-1)
:negative