Elixir In Action v0.1.0 Chapter3.FileHelper

This module explain how to use high-order functions (Enum and Stream) to iterate over enumerables

Book section: 3.4

Functions

large_lines!(path)

Specs

large_lines!(String.t()) :: [String.t()]

Get just the lines above 80 characters from a file

Book section: 3.4.5

Examples

iex> Chapter3.FileHelper.large_lines!("./test/chapters/.resources/text_file.txt")
["this is a very very very very very very very very very very very very very long line"]

lines_lengths!(path)

Specs

lines_lengths!(String.t()) :: [non_neg_integer()]

Get the lines length of a file

Book section: 3.4.5

Examples

iex> Chapter3.FileHelper.lines_lengths!("./test/chapters/.resources/text_file.txt")
[1,3,5,7,9,84,35,23]

longest_line!(path)

Specs

longest_line!(String.t()) :: String.t()

Get the longest line of a file

Book section: 3.4.5

Examples

iex> Chapter3.FileHelper.longest_line!("./test/chapters/.resources/text_file.txt")
"this is a very very very very very very very very very very very very very long line"

longest_line_length!(path)

Specs

longest_line_length!(String.t()) :: non_neg_integer()

Get the longest line length of a file

Book section: 3.4.5

Examples

iex> Chapter3.FileHelper.longest_line_length!("./test/chapters/.resources/text_file.txt")
84

words_per_line!(path)

Specs

words_per_line!(String.t()) :: [non_neg_integer()]

Get the the number of words per line of a file

Book section: 3.4.5

Examples

iex> Chapter3.FileHelper.words_per_line!("./test/chapters/.resources/text_file.txt")
[1,2,3,4,5,18,18,12]