Elixir In Action v0.1.0 Chapter3.ListHelper

This module explains how pattern macthing and multiclause functions are powerfull tools to build recursive functions and how use tail and non-tail recursions to build iterations

Book section: 3.3 and 3.4

Functions

no_tail_len(list)

Specs

no_tail_len(list()) :: number()

Get the length of a list using non tail recursive function call

Book section: 3.4.1

This exercises shows how to built iterations using recursive functions

And shows how to use the list notation of [head | tails]

Examples

iex> Chapter3.ListHelper.no_tail_len([2,3,5,10])
4

iex> Chapter3.ListHelper.no_tail_len([0,0,33,2,7])
5

iex> Chapter3.ListHelper.no_tail_len([])
0

non_tail_positives(list)

Specs

non_tail_positives(list()) :: list()

Get a list with just the positives integer on a list using non tail recursive call

Book section: 3.4.2

This exercises shows how to build iterations using non tail recursive functions

Examples

iex> Chapter3.ListHelper.non_tail_positives([0,5,7,-3,8,-2,-1])
[5,7,8]

iex> Chapter3.ListHelper.non_tail_positives([3,2,1])
[3,2,1]

iex> Chapter3.ListHelper.non_tail_positives([-7,-10])
[]

non_tail_range(from, to)

Specs

non_tail_range(number(), number()) :: list()

Get a list of integers between 2 numbers using non tail recursive call

Book section: 3.4.2

This exercises shows how to build iterations using non tail recursive functions

Examples

iex> Chapter3.ListHelper.non_tail_range(0,5)
[0,1,2,3,4,5]

iex> Chapter3.ListHelper.non_tail_range(7,7)
[7]

iex> Chapter3.ListHelper.non_tail_range(7,10)
[7,8,9,10]

sum(list)

Specs

sum(list()) :: number()

Sum all elements on a list

Book section: 3.3.1 | 3.4.1

This exercises shows how to use multiclause functions to build recursive functions more easily

And shows how to use the list notation of [head | tails]

Examples

iex> Chapter3.ListHelper.sum([2,3,5,10])
20

iex> Chapter3.ListHelper.sum([0,0,33,2,7])
42

iex> Chapter3.ListHelper.sum([])
0

tail_len(list)

Specs

tail_len(list()) :: number()

Get the length of a list using tail recursive function call

Book section: 3.4.2

This exercises shows how to build iterations using tail recursive functions

And shows how to use the list notation of [head | tails]

Examples

iex> Chapter3.ListHelper.tail_len([2,3,5,10])
4

iex> Chapter3.ListHelper.tail_len([0,0,33,2,7])
5

iex> Chapter3.ListHelper.tail_len([])
0

tail_positives(list)

Specs

tail_positives([number()]) :: [non_neg_integer()]

Get a list with just the positives integer on a list using tail recursive call

Book section: 3.4.2

This exercises shows how to build iterations using tail recursive functions

Examples

iex> Chapter3.ListHelper.tail_positives([0,5,7,-3,8,-2,-1])
[5,7,8]

iex> Chapter3.ListHelper.tail_positives([3,2,1])
[3,2,1]

iex> Chapter3.ListHelper.tail_positives([-7,-10])
[]

tail_range(from, to)

Specs

tail_range(number(), number()) :: list()

Get a list of integers between 2 numbers using tail recursive call

Book section: 3.4.2

This exercises shows how to build iterations using tail recursive functions

The logic was made inverse because of performance, since lists are O(1) to insert on the start and O(n) to insert on the end its better to build the list from end to start

Examples

iex> Chapter3.ListHelper.tail_range(0,5)
[0,1,2,3,4,5]

iex> Chapter3.ListHelper.tail_range(7,7)
[7]

iex> Chapter3.ListHelper.tail_range(7,10)
[7,8,9,10]