Elixir In Action v0.1.0 Chapter3.MapHelper

This module explain how to use the with special form

Book section: 3.3

Functions

normalize_user!(user)

Specs

normalize_user!(map()) ::
  {:error, String.t()}
  | {:ok, %{email: String.t(), login: String.t(), password: String.t()}}

Normalize a map to a map containing :login, :email and :password keys

Book section: 3.3.3

This exercises shows how to use with macro and pattern matching to achieve more elegant code.

Without with a chain of case or if would be necessary.

Examples

iex> Chapter3.MapHelper.normalize_user!(%{"wrong" => "data"})
{:error, "login missing"}

iex> Chapter3.MapHelper.normalize_user!(%{"login" => "example"})
{:error, "email missing"}

iex> Chapter3.MapHelper.normalize_user!(%{"login" => "example", "email" => "example@email.com"})
{:error, "password missing"}

iex> Chapter3.MapHelper.normalize_user!(%{"login" => "example", "email" => "example@email.com", "password" => "secret", "other" => "data"})
{:ok, %{login: "example", email: "example@email.com", password: "secret"}}