Elixir In Action v0.1.0 Chapter4.Fraction

This module explains how to build higher level data structures with built-in types and how to use struct and apply pattern matching with it

Book Section 4.1

Types

fraction()

Specs

fraction() :: %Chapter4.Fraction{denominator: number(), numerator: number()}

In this example, the struct fraction is built with the built-in type map. Because of the usage of struct it is slight different of a conventional map, but most functions from the Map module will work with it

Functions

add(fraction1, fraction2)

Specs

add(fraction(), fraction()) :: fraction()

Add two fractions and return the result

Book section: 4.1.4

This function shows how to use pattern matching with structs

Examples

iex> Chapter4.Fraction.add(Chapter4.Fraction.new(1,2), Chapter4.Fraction.new(1,4))
%Chapter4.Fraction{numerator: 6, denominator: 8}

iex> Chapter4.Fraction.add(Chapter4.Fraction.new(3,2), Chapter4.Fraction.new(1,4))
%Chapter4.Fraction{numerator: 14, denominator: 8}

new(numerator, denominator)

Specs

new(integer(), integer()) :: fraction()

Create a new fraction data abstraction

Book section: 4.1.4

This function shows how to build constructors-like behaviour on elixir

Examples

iex> Chapter4.Fraction.new(1,2)
%Chapter4.Fraction{numerator: 1, denominator: 2}

iex> Chapter4.Fraction.new(5,3)
%Chapter4.Fraction{numerator: 5, denominator: 3}

value(fraction)

Specs

value(fraction()) :: float()

Get the float equivalent value of a fraction

Book section: 4.1.4

This function shows how to use pattern matching with structs

Examples

iex> Chapter4.Fraction.new(1,2) |> Chapter4.Fraction.value()
0.5

iex> Chapter4.Fraction.value(%Chapter4.Fraction{numerator: 3, denominator: 4})
0.75