Elixir In Action v0.1.0 Chapter5.QueryHelper

This module explains how to spawn a new process and receive messages from it

Book section: 5.2

Functions

execute_query(query_text)

Specs

execute_query(String.t()) :: <<_::56, _::_*8>>

Simulate a database call with 2 seconds delay

Book section: 5.2.1

execute_query_on_another_process(query_text)

Specs

execute_query_on_another_process(String.t()) :: pid()

Spawn a new process to executes a query

self/1 is used to get the caller pid wich will be send to the new process

self/1 can should not be called inside the spawn/1 function

A deep copy of every data passed to spawn/1 is send to the new process

Book section: 5.2.1

get_some_result_from_inbox()

Get the first query result from the process inbox

This function relies on patten matching to distinguish between query result messages and other possible messages on the inbox

To prevent being stuck if no message has arrived, the clause after is used. Kind of a timeout

Book section: 5.2.2

run_queries_async(queries_list)

Specs

run_queries_async(list()) :: [pid()]

Run a list of queries asynchronously

This function always takes 2 seconds to execute

Although, it does not return the queries results, it returns a list of pids

Book section: 5.2.2

run_queries_sync(queries_list)

Specs

run_queries_sync(list()) :: list()

Run a list of queries synchronously

This function takes 2 * length(list) seconds to execute

Book section: 5.2.1