Elixir In Action v0.1.0 Chapter6.KeyValueStore

This module explains how to build a module that relies on GenServer

The basic approach is similiar to the Chapter6.MyKeyValueStore

This callback module must implement a set of functions in order to be compatible with GenServer

Book section: 6.2

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

get(server_pid, key)

Get the value referenced by a specific key synchronously

If a function should be running synchronously, it must use the GenServer.call/2 function

Book section: 6.2.3

Examples

iex> {:ok, pid} = Chapter6.KeyValueStore.start()
iex> Chapter6.KeyValueStore.put(pid, :first, "value")
iex> Chapter6.KeyValueStore.get(pid, :first)
"value"

iex> {:ok, pid} = Chapter6.KeyValueStore.start()
iex> Chapter6.KeyValueStore.get(pid, :first)
nil

handle_info(msg, state)

Handle generic custom calls to the server

Any message that arrives to the server that do not fall into one of the 2 categories, cast or call will be handled by this function

Book section: 6.2.4

init(args)

Initialize the server

The function :timer.send_interval is used to send a periodically request to the server

Book section: 6.2.4

put(server_pid, key, value)

Add a key/value pair asynchronously

If a function should be running asynchronously, it must use the GenServer.cast/2 function

Book section: 6.2.3

Examples

iex> {:ok, pid} = Chapter6.KeyValueStore.start()
iex> Chapter6.KeyValueStore.put(pid, :first, "value")
iex> Chapter6.KeyValueStore.get(pid, :first)
"value"

start()

Starts the server process

This function relies in GenServer.start/3

Book section: 6.2.2

Examples

iex> {:ok, pid} = Chapter6.KeyValueStore.start()
iex> is_pid(pid)
true