Elixir In Action v0.1.0 Chapter6.MyKeyValueStore

This module explains how to build a concrete module that relies on a generic server implementation

The concrete module relies on functions such as Chapter6.MyServerProcess.start/1, Chapter6.MyServerProcess.cast/2 and Chapter6.MyServerProcess.call/2 to implement a server process with no boilerplate

The clarity of the code using a GenServer abstraction is way higher

Book section: 6.1

Functions

get(pid, key)

Specs

get(pid(), any()) :: any()

Get the value referenced by a specific key synchronously

If a function should be running synchronously, it must use the Chapter6.MyServerProcess.call/2 function

Book section: 6.1.3

Examples

iex> pid = Chapter6.MyKeyValueStore.start()
iex> Chapter6.MyKeyValueStore.put(pid, :first, "value")
iex> Chapter6.MyKeyValueStore.get(pid, :first)
"value"

handle_call(arg, state)

Specs

handle_call({:get, any()}, map()) :: {any(), map()}

Handle a sync get message

The return of the handle_cast functions implementations should always be in the same format that the abstraction expects {result, new_state}

Book section: 6.1.3

Examples

iex> Chapter6.MyKeyValueStore.handle_call({:get, :key} , %{key: "value"})
{"value", %{key: "value"}}

iex> Chapter6.MyKeyValueStore.handle_call({:get, :key} , %{})
{nil, %{}}

handle_cast(arg, state)

Specs

handle_cast({:put, any(), any()}, map()) :: map()

Handle an async put message

The return of the handle_cast functions implementations should always be the new state value

Book section: 6.1.4

Examples

iex> Chapter6.MyKeyValueStore.handle_cast({:put, :key, "value"}, %{})
%{key: "value"}

iex> Chapter6.MyKeyValueStore.handle_cast({:put, :key, "value"}, %{other_key: "other_value"})
%{key: "value", other_key: "other_value"}

init()

Specs

init() :: %{}

Return the initial state of the server

This function is used inside the Chapter6.MyServerProcess.start/1 implementation to set the initial state of the server process

Book section: 6.1.3

Examples

iex> Chapter6.MyKeyValueStore.init()
%{}

put(pid, key, value)

Specs

put(pid(), any(), any()) :: any()

Add a key/value pair asynchronously

If a function should be running asynchronously, it must use the Chapter6.MyServerProcess.cast/2 function

Book section: 6.1.3

Examples

iex> pid = Chapter6.MyKeyValueStore.start()
iex> Chapter6.MyKeyValueStore.put(pid, :first, "value")
iex> Chapter6.MyKeyValueStore.get(pid, :first)
"value"

start()

Starts the server process

This function relies in Chapter6.MyServerProcess.start/1

Book section: 6.1.3

Examples

iex> pid = Chapter6.MyKeyValueStore.start()
iex> is_pid(pid)
true