Elixir In Action v0.1.0 Chapter6.MyServerProcess

This module explains how to build a generic server process that abstracts implementation details to build a server process

A concrete module will be provided by dependency injection to fill the specific implementation but all the flow of control will be managed by this process

This helps the concrete modules to be focused on the implementation details of what is relevant for it, making it more clean and easy to understand

To this method work, the concrete module has to implement a well defined set of functions very similar to a interface in OO

Book section: 6.1

Functions

call(server_pid, request)

Specs

call(pid(), any()) :: any()

Make a call to the server process and await for it response

This function will handle all the requests to the concrete server that the client want to know the result

The concrete function must call this function always that a interface function should run synchronously. eg.this function

Book section: 6.1.2

Examples

iex> pid = Chapter6.MyStackServer.start() iex> Chapter6.MyStackServer.put(pid, "first") iex> Chapter6.MyServerProcess.call(pid, {:get}) "first"

iex> pid = Chapter6.MyStackServer.start() iex> Chapter6.MyServerProcess.call(pid, {:get}) :stack_is_empty

cast(server_pid, request)

Specs

cast(pid(), any()) :: any()

Make a call to the server process and do not wait for it response

This function will handle all the requests to the concrete server that the client do not want to know the result

The concrete function must call this function always that a interface function should run asynchronously. eg.this function

Book section: 6.1.4

Examples

iex> pid = Chapter6.MyStackServer.start() iex> Chapter6.MyServerProcess.cast(pid, {:put, "first"}) {:cast, {:put, "first"}}

start(callback_module)

Specs

start(module()) :: pid()

Starts the injected module as a server process

The injected module must implement init/0 function, that return the initial state of the server process

The injected module is passed as parameter to the loop/2 function

Book section: 6.1.1

Examples

iex> pid = Chapter6.MyServerProcess.start(Chapter6.MyStackServer)
iex> is_pid(pid)
true